+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All versions prior to 0.0.9 are untracked.
 
 ## [Unreleased]
 
+### Added
+
+* pip-audit now includes vulnerability aliases when `--format=json` is used
+
 ## [2.6.3]
 
 ### Fixed